requirements.txt
pandas
numpy
sqlalchemy
pymysql
matplotlib
openpyxl
jupyter
python-dotenv